+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.0.0 (unreleased)
+  * APNs error detection is now performed asynchronously, 'check_for_errors' is therefore deprecated.
   * Deprecated attributes_for_device accessors. Use data instead.
   * Fix signal handling to work with Ruby 2.x. (#40).
   * You no longer need to signal HUP after creating a new app, they will be loaded automatically for you.